tests/wpt/web-platform-tests/XMLHttpRequest/open-url-encoding.htm

<!DOCTYPE html>
<html>
  <head>
    <meta charset=windows-1252>
    <title>XMLHttpRequest: open() - URL encoding</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-open()-method" data-tested-assertations="following::ol/li[7] following::ol/li[14]/ul/li[2]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/content.py?\u00DF", false) // This is the German "eszett" character
        client.send()
        assert_equals(client.getResponseHeader("x-request-query"), "%C3%9F")
      }, "percent encode characters");
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "resources/content.py?\uD83D", false)
        client.send()
        assert_equals(client.getResponseHeader("x-request-query"), "%EF%BF%BD")
      }, "lone surrogate should return U+FFFD");
    </script>
  </body>
</html>
